u32 -> i64

+++ src/groups/groups.rs
@@ -17,12 +17,12 @@ use super::{permissions::GroupPermissions, roles::GroupRole};
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Group {
-    pub id: u32,
+    pub id: i64,
     pub name: String,
     pub description: String,
     pub owner: MinimalGroupUser,
     pub shout: Option<GroupShout>,
-    pub member_count: Option<u32>,
+    pub member_count: Option<i64>,
     pub is_builders_club_only: bool,
     pub public_entry_allowed: bool,
     pub is_locked: Option<bool>,
@@ -42,7 +42,7 @@ pub struct GroupShout {
 ///
 /// # Error codes
 /// - 1: Group is invalid or does not exist.
-pub async fn group_by_id(jar: &RequestJar, group_id: u32) -> Result<Group, Box<Error>> {
+pub async fn group_by_id(jar: &RequestJar, group_id: i64) -> Result<Group, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/{}", group_id);
     let response = jar.get_json::<Group>(&url).await?;
     Ok(response)
@@ -121,9 +121,9 @@ pub enum GroupAuditLogActionType {
 /// - 23: Insufficient permissions to complete the request.
 pub async fn audit_log(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     limit: PageLimit,
-    user_id: Option<u32>,
+    user_id: Option<i64>,
     sort_order: Option<SortOrder>,
     //cursor: Option<String>,
 ) -> Result<Vec<GroupAuditLogEntry>, Box<Error>> {
@@ -152,7 +152,7 @@ pub struct GroupNameHistoryEntry {
 /// - 23: Insufficient permissions to complete the request.
 pub async fn name_history(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     limit: PageLimit,
     sort_order: Option<SortOrder>,
     //cursor: Option<String>,
@@ -183,7 +183,7 @@ pub struct GroupSettings {
 /// # Error codes
 /// - 1: Group is invalid or does not exist.
 /// - 23: Insufficient permissions to complete the request.
-pub async fn settings(jar: &RequestJar, group_id: u32) -> Result<GroupSettings, Box<Error>> {
+pub async fn settings(jar: &RequestJar, group_id: i64) -> Result<GroupSettings, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/{}/settings", group_id);
     let response = jar.get_json::<GroupSettings>(&url).await?;
     Ok(response)
@@ -210,7 +210,7 @@ pub struct GroupSettingsUpdateResponse {}
 /// - 31: Service is currently unavailable.
 pub async fn update_settings(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     request: GroupSettingsUpdateRequest,
 ) -> Result<GroupSettingsUpdateResponse, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/{}/settings", group_id);
@@ -226,7 +226,7 @@ pub async fn update_settings(
 #[serde(rename_all = "camelCase")]
 pub struct GroupComplianceItem {
     pub can_view_group: bool,
-    pub group_id: u32,
+    pub group_id: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
@@ -238,7 +238,7 @@ pub struct GroupComplianceResponse {
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupComplianceRequest {
-    pub group_ids: Vec<u32>,
+    pub group_ids: Vec<i64>,
 }
 
 /// Gets group policy info used for compliance
@@ -247,7 +247,7 @@ pub struct GroupComplianceRequest {
 /// - 2: Ids could not be parsed from request.
 pub async fn compliance(
     jar: &RequestJar,
-    group_ids: Vec<u32>,
+    group_ids: Vec<i64>,
 ) -> Result<GroupComplianceResponse, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/policies");
     let request = GroupComplianceRequest { group_ids };
@@ -277,7 +277,7 @@ pub struct NewDescriptionResponse {
 /// - 29: Your group description was empty.
 pub async fn update_description(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     description: String,
 ) -> Result<NewDescriptionResponse, Box<Error>> {
     let url = format!(
@@ -314,7 +314,7 @@ pub struct NewNameResponse {
 /// - 29: Your group description was empty.
 pub async fn update_name(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     name: String,
 ) -> Result<NewNameResponse, Box<Error>> {
     let url = format!(
@@ -335,7 +335,7 @@ pub async fn update_name(
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupMembership {
-    pub group_id: u32,
+    pub group_id: i64,
     pub is_primary: bool,
     pub is_pending_join: bool,
     pub group_role: Option<GroupMembershipUserRole>,
@@ -357,7 +357,7 @@ pub struct GroupMembershipUserRole {
 ///
 /// # Error codes
 /// - 1: Group is invalid or does not exist.
-pub async fn membership(jar: &RequestJar, group_id: u32) -> Result<GroupMembership, Box<Error>> {
+pub async fn membership(jar: &RequestJar, group_id: i64) -> Result<GroupMembership, Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/membership",
         group_id
@@ -372,7 +372,7 @@ pub async fn membership(jar: &RequestJar, group_id: u32) -> Result<GroupMembersh
 /// - 1: The group is invalid or does not exist.
 pub async fn members(
     jar: &RequestJar,
-    group_id: u32,
+    group_id: i64,
     limit: PageLimit,
     sort_order: Option<SortOrder>,
 ) -> Result<Vec<GroupMembershipUserRole>, Box<Error>> {
@@ -447,7 +447,7 @@ pub struct UserMembershipsGroupItem {
 /// - 3: The user is invalid or does not exist.
 pub async fn user_memberships(
     jar: &RequestJar,
-    user_id: u32,
+    user_id: i64,
 ) -> Result<Vec<UserMembershipsGroupItem>, Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/users/{}/groups/roles",
@@ -462,7 +462,7 @@ pub async fn user_memberships(
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupOwnershipChangeRequest {
-    pub user_id: u32,
+    pub user_id: i64,
 }
 
 /// Changes the owner of a group
@@ -474,7 +474,7 @@ pub struct GroupOwnershipChangeRequest {
 /// - 16: The user does not have the necessary level of premium membership.
 /// - 17: You are not authorized to change the owner of this group.
 /// - 25: 2-Step Verification is required to make further transactions. Go to Settings > Security to complete 2-Step Verification.
-pub async fn change_owner(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
+pub async fn change_owner(jar: &RequestJar, group_id: i64, user_id: i64) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/change-owner",
         group_id
@@ -492,7 +492,7 @@ pub async fn change_owner(jar: &RequestJar, group_id: u32, user_id: u32) -> Resu
 /// - 12: This group already has an owner.
 /// - 13: Too many attempts to claim groups. Please try again later.
 /// - 18: The operation is temporarily unavailable. Please try again later.
-pub async fn claim_ownership(jar: &RequestJar, group_id: u32) -> Result<(), Box<Error>> {
+pub async fn claim_ownership(jar: &RequestJar, group_id: i64) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/claim-ownership",
         group_id
