+++ src/groups/join_requests.rs
@@ -14,7 +14,7 @@ pub mod join_requests {
     #[derive(Debug, Serialize, Deserialize, Clone, Type)]
     #[serde(rename_all = "camelCase")]
     pub struct BatchRequest {
-        pub user_ids: Vec<u32>,
+        pub user_ids: Vec<i64>,
     }
 
     #[derive(Debug, Serialize, Deserialize, Clone, Type)]
@@ -32,7 +32,7 @@ pub mod join_requests {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn get(
         jar: &RequestJar,
-        group_id: u32,
+        group_id: i64,
         limit: PageLimit,
     ) -> Result<Vec<JoinRequest>, Box<Error>> {
         let url = format!(
@@ -55,8 +55,8 @@ pub mod join_requests {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn accept(
         jar: &RequestJar,
-        group_id: u32,
-        user_ids: Vec<u32>,
+        group_id: i64,
+        user_ids: Vec<i64>,
     ) -> Result<(), Box<Error>> {
         let url = format!(
             "https://groups.roblox.com/v1/groups/{}/join-requests",
@@ -75,8 +75,8 @@ pub mod join_requests {
     #[allow(unused)] // It appears to be a bug in the checker, this kinda fixes it.
     pub async fn decline(
         jar: &RequestJar,
-        group_id: u32,
-        user_ids: Vec<u32>,
+        group_id: i64,
+        user_ids: Vec<i64>,
     ) -> Result<(), Box<Error>> {
         let url = format!(
             "https://groups.roblox.com/v1/groups/{}/join-requests",
@@ -100,7 +100,7 @@ pub mod join_request {
     #[derive(Debug, Serialize, Deserialize, Clone, Type)]
     #[serde(rename_all = "camelCase")]
     pub struct BatchRequest {
-        pub user_ids: Vec<u32>,
+        pub user_ids: Vec<i64>,
     }
 
     #[derive(Debug, Serialize, Deserialize, Clone, Type)]
@@ -117,8 +117,8 @@ pub mod join_request {
     /// - 19: You have insufficient permissions for this request.
     pub async fn get(
         jar: &RequestJar,
-        group_id: u32,
-        user_id: u32,
+        group_id: i64,
+        user_id: i64,
     ) -> Result<JoinRequest, Box<Error>> {
         let url = format!(
             "https://groups.roblox.com/v1/groups/{}/join-requests/users/{}",
@@ -137,7 +137,7 @@ pub mod join_request {
     /// - 18: The operation is temporarily unavailable. Please try again later.
     /// - 19: You have insufficient permissions for this request.
     /// - 20: The group join request is invalid.
-    pub async fn accept(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
+    pub async fn accept(jar: &RequestJar, group_id: i64, user_id: i64) -> Result<(), Box<Error>> {
         let url = format!(
             "https://groups.roblox.com/v1/groups/{}/join-requests/users/{}",
             group_id, user_id
@@ -151,7 +151,7 @@ pub mod join_request {
     /// # Error codes
     /// - 3: The user is invalid or does not exist.
     /// - 4: You do not have permission to manage this member.
-    pub async fn decline(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
+    pub async fn decline(jar: &RequestJar, group_id: i64, user_id: i64) -> Result<(), Box<Error>> {
         let url = format!(
             "https://groups.roblox.com/v1/groups/{}/join-requests/users/{}",
             group_id, user_id
