+++ src/groups/metadata.rs
@@ -17,36 +17,36 @@ pub struct GroupConfigMetadata {
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupConfigurationMetadata {
-    pub name_max_length: u32,
-    pub description_max_length: u32,
-    pub icon_max_file_size_mb: u32,
-    pub cost: u32,
+    pub name_max_length: i64,
+    pub description_max_length: i64,
+    pub icon_max_file_size_mb: i64,
+    pub cost: i64,
     pub is_using_two_step_webview_component: bool,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRecurringPayoutsConfigurationMetadata {
-    pub max_payout_partners: u32,
+    pub max_payout_partners: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRoleConfigurationMetadata {
-    pub name_max_length: u32,
-    pub description_max_length: u32,
-    pub limit: u32,
-    pub cost: u32,
-    pub min_rank: u32,
-    pub max_rank: u32,
+    pub name_max_length: i64,
+    pub description_max_length: i64,
+    pub limit: i64,
+    pub cost: i64,
+    pub min_rank: i64,
+    pub max_rank: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupNameChangeConfigurationMetadata {
-    pub cost: u32,
-    pub cooldown_in_days: u32,
-    pub ownership_cooldown_in_days: u32,
+    pub cost: i64,
+    pub cooldown_in_days: i64,
+    pub ownership_cooldown_in_days: i64,
 }
 
 /// Gets the metadata for group configuration
@@ -59,12 +59,12 @@ pub async fn config_metadata(jar: &RequestJar) -> Result<GroupConfigMetadata, Bo
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupMetadata {
-    pub group_limit: u32,
-    pub current_group_count: u32,
-    pub group_status_max_length: u32,
-    pub group_post_max_length: u32,
+    pub group_limit: i64,
+    pub current_group_count: i64,
+    pub group_status_max_length: i64,
+    pub group_post_max_length: i64,
     pub is_group_wall_notifications_enabled: bool,
-    pub group_wall_notifications_subscribe_interval_in_milliseconds: u32,
+    pub group_wall_notifications_subscribe_interval_in_milliseconds: i64,
     pub are_profile_groups_hidden: bool,
     pub is_group_details_policy_enabled: bool,
     pub show_previous_group_names: bool,
