+++ src/groups/roles.rs
@@ -16,17 +16,17 @@ use super::user_memberships;
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRole {
-    pub id: u32,
+    pub id: i64,
     pub name: String,
     pub description: Option<String>,
-    pub rank: u32,
-    pub member_count: Option<u32>,
+    pub rank: i64,
+    pub member_count: Option<i64>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRoleResponse {
-    pub group_id: u32,
+    pub group_id: i64,
     pub roles: Vec<GroupRole>,
 }
 
@@ -34,7 +34,7 @@ pub struct GroupRoleResponse {
 ///
 /// # Error codes
 /// - 1: The group is invalid or does not exist.
-pub async fn roles(jar: &RequestJar, group_id: u32) -> Result<Vec<GroupRole>, Box<Error>> {
+pub async fn roles(jar: &RequestJar, group_id: i64) -> Result<Vec<GroupRole>, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/{}/roles", group_id);
     let response = jar.get_json::<GroupRoleResponse>(&url).await?;
     Ok(response.roles)
@@ -47,8 +47,8 @@ pub async fn roles(jar: &RequestJar, group_id: u32) -> Result<Vec<GroupRole>, Bo
 /// - 2: The roleset is invalid or does not exist.
 pub async fn users_on_role(
     jar: &RequestJar,
-    group_id: u32,
-    role_id: u32,
+    group_id: i64,
+    role_id: i64,
     limit: PageLimit,
     sort_order: Option<SortOrder>,
 ) -> Result<Vec<MinimalGroupUser>, Box<Error>> {
@@ -72,8 +72,8 @@ pub async fn users_on_role(
 /// - 200: The user is not in the group.
 pub async fn user_role(
     jar: &RequestJar,
-    group_id: u32,
-    user_id: u32,
+    group_id: i64,
+    user_id: i64,
 ) -> Result<GroupRole, Box<Error>> {
     let roles = user_memberships(jar, user_id).await?;
 
@@ -101,7 +101,7 @@ pub async fn user_role(
 /// - 2: Too many ids in request.
 pub async fn roles_by_id(
     jar: &RequestJar,
-    role_ids: Vec<u32>,
+    role_ids: Vec<i64>,
 ) -> Result<Vec<GroupRole>, Box<Error>> {
     let string_ids = role_ids.iter().map(|id| id.to_string()).collect::<Vec<_>>();
     let url = format!(
